pyproject.toml
[project]
name = "day9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "bson>=0.5.10",
    "celery[redis]>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "openai>=1.99.4",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=5.2.1",
    "ruff>=0.12.7",
    "scalar-fastapi>=1.2.3",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 120
lint.select = ["E", "W", "F", "I"]
exclude = ["alembic"]
